backend/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "textbook-ingestion"
version = "0.1.0"
description = "Website ingestion pipeline for Docusaurus textbooks with vector storage"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Physical AI Robotics", email = "info@example.com"}
]
keywords = ["ingestion", "embeddings", "qdrant", "cohere", "vector-search"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "langchain-text-splitters>=0.0.1",
    "tiktoken>=0.5.0",
    "cohere>=5.0.0",
    "qdrant-client>=2.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
]

[project.urls]
Homepage = "https://github.com/example/physical-ai-robotics-textbook"
Documentation = "https://github.com/example/physical-ai-robotics-textbook/blob/main/backend/README.md"
Repository = "https://github.com/example/physical-ai-robotics-textbook"
Issues = "https://github.com/example/physical-ai-robotics-textbook/issues"

[project.scripts]
textbook-ingest = "ingestion.main:main"

[tool.pytest.ini_options]
minversion = "7.4"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers"
markers = [
    "skip: marks tests as skipped (deselect with '-m \"not skip\"')",
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "live_api: marks tests that require live API access",
]

[tool.coverage.run]
source = ["ingestion"]
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*Protocol.*:",
]
precision = 2
skip_covered = false

[tool.isort]
profile = "black"
line_length = 100
skip_glob = ["*/migrations/*"]
known_first_party = ["ingestion"]

[tool.black]
line-length = 100
target-version = ["py310"]
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | __pycache__
  | \.pytest_cache
  | \.eggs
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
